Moved ManagedSensor to its own class
